QmlProfiler: Always use quint16 for port number

+++ src/plugins/qmlprofiler/abstractqmlprofilerrunner.h
@@ -49,7 +49,7 @@ public:
     virtual void start() = 0;
     virtual void stop() = 0;
 
-    virtual int debugPort() const = 0;
+    virtual quint16 debugPort() const = 0;
 
 signals:
     void started();
